toolkit/components/satchel/test/test_form_autocomplete_validation_at_input_event.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for validation has been done before "input" event</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>
  input:invalid {
    background-color: red;
  }
  </style>
</head>
<body>
<p id="display"></p>

<div id="content">
  <form>
    <input type="text" name="field1" pattern="[0-9]{4}">
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(async () => {
  function promiseInitFormHistory() {
    return new Promise(resolve => {
      updateFormHistory([
        { op: "remove" },
        { op: "add", fieldname: "field1", value: "1234" },
        { op: "add", fieldname: "field1", value: "12345" },
      ], resolve);
    });
  }
  await promiseInitFormHistory();

  let input = document.querySelector("input[name=field1]");
  input.focus();

  let resolver;
  function onPopup() {
    if (resolver) {
      resolver();
    }
  }
  registerPopupShownListener(onPopup);

  function promiseOpenPopup() {
    return new Promise(resolve => {
      resolver = resolve;
    });
  }
  let waitForOpenPopup = promiseOpenPopup();
  let inputFired = false;
  ok(true, "Typing first character to open popup...");
  input.addEventListener("input", (aEvent) => {
    inputFired = true;
    is(aEvent.inputType, "insertText", 'Typing "1" should cause an "input" event whose inputType is "insertText"');
    is(input.validity.valid, false, 'Typing "1" should mark it as "invalid"');
    is(input.matches(":invalid"), true, 'Typing "1" should cause matching with ":invalid" pseudo-class');
  }, {once: true, capture: true});
  synthesizeKey("1");
  ok(inputFired, 'Typing "1" should cause one "input" event');
  await waitForOpenPopup;

  inputFired = false;
  input.addEventListener("input", (aEvent) => {
    inputFired = true;
    is(aEvent.inputType, "insertReplacementText", 'Selecting valid value should cause an "input" event whose inputType is "insertReplacementText"');
    is(input.validity.valid, true, 'Selecting valid value should mark it as "valid"');
    is(input.matches(":invalid"), false, 'Selecting valid value should cause not matching with ":invalid" pseudo-class');
  }, {once:true, capture:true});
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  ok(inputFired, 'Selecting valid item should cause an "input" event');

  ok(true, "Reopening popup...");
  waitForOpenPopup = promiseOpenPopup();
  synthesizeKey("KEY_ArrowDown");
  await waitForOpenPopup;

  inputFired = false;
  input.addEventListener("input", (aEvent) => {
    inputFired = true;
    is(aEvent.inputType, "insertReplacementText", 'Selecting invalid value should cause an "input" event whose inputType is "insertReplacementText"');
    is(input.validity.valid, false, 'Selecting invalid value should mark it as "invalid"');
    is(input.matches(":invalid"), true, 'Selecting invalid value should cause matching with ":invalid" pseudo-class');
  }, {once:true, capture:true});
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  ok(inputFired, 'Selecting invalid item should cause an "input" event');

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
